Refactor codegen out of transform classes and into classes specifically for codegen.
