+++ SharpGen/Generator/ICodeGenerator.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
 using SharpGen.Model;
 using System;
 using System.Collections.Generic;
@@ -8,9 +7,9 @@ using System.Text;
 namespace SharpGen.Generator
 {
     interface ICodeGenerator<in TCsElement, out TSyntax>
-        where TCsElement: CsBase
-        where TSyntax: SyntaxNode
+        where TCsElement : CsBase
+        where TSyntax : SyntaxNode
     {
-        IEnumerable<TSyntax> GenerateCode(TCsElement csElement);
+        TSyntax GenerateCode(TCsElement csElement);
     }
 }
